requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
astunparse==1.6.3
boto3==1.34.49
botocore==1.34.49
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.4
fastapi==0.109.2
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.28.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.62.0
h11==0.14.0
h5py==3.10.0
idna==3.6
jax==0.4.24
jmespath==1.0.1
keras==2.12.0
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
ml-dtypes==0.2.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pdfminer.six==20221105
pdfplumber==0.10.4
pillow==10.2.0
pinecone-client==3.0.3
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
pypdfium2==4.27.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.10.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.0.7
uvicorn==0.27.1
Werkzeug==3.0.1
wrapt==1.14.1
